Automatic generation of virtual waypoint based on wind direction.
The firmware generates three virtual waypoints between the last mission WP and the landing WP based on the wind direction. The purpose is to have the UAV approaching the landing WP always with at least 90 degree of heading respect to the wind direction, before the transition between fixed-wing mode and quadrotor-mode.