Altitude during virtual waypoint progressively reach the landing waypoint altitude.
The altitude during the virtual waypoint is calculated progressively based on the altitude of the last mission waypoint and the landing waypoint. The goal is to set the altitude of last virtual waypoint equal to the landing waypoint with the UAV flying a limited speed before the transition to quadrotor mode.